+++ js/menu.js
@@ -8,7 +8,6 @@ var workspace = require("./workspace");
 var pipelineGate = require("./pipeline_gate");
 var dialogs = require("./dialogs");
 var appLogToggle = require("./app_log_toggle");
-var ioFairshareSettings = require("./io_fairshare_settings");
 
 var WORKSPACE_MENU = "./workspace_menu.html";
 
@@ -207,7 +206,6 @@ pageInit.onReady(function () {
 
 	bindOpenProject();
 	appLogToggle.bindAppLogToggle(toggleAppLogBtn);
-	ioFairshareSettings.bindNetworkSharing("ioFairsharePanel");
 	if (checkUpdatesBtn) {
 		checkUpdatesBtn.addEventListener("click", function () {
 			ipc.send("checkForUpdates", []);